[COVERAGE PIPE] small fix: move parsePipeOptions() to pipe_utils.js
